zsh: Improve vcs_info_async prompt
Improved detection of submodules by looking for ".git" and ".svn"
files/folders. If these files/folders exist, a refresh is immediately
triggered.

Additionally, replaced the default ellipsis used by vcs_info with a
customizable string, such as to shorten it to a single unicode
character.

Last but not least, changed the default 'working' symbol from an
ellipsis to a question mark, to avoid confusion with the truncated
branch name ellipsis used by git.