vim: Fix recursive mapping on TAB/CR
When using the TAB/CR mappings, sometimes vim would freeze when
expand_tabs=0, forcing the user to press Ctrl-C to cancel the mapping.

This was due to accidentally using *map instead of *noremap for the
TAB/CR mappings in plugins/autocomplete.vim, which caused the symbol
"\<Tab>" output by GetTabCharacters() to trigger the mapping
recursively.